refactor(dscp): use FQN versioned proto package
Migrate the dscp module gRPC contract to a fully-qualified, versioned proto
package (modules.dscp.controlplane.dscppb.v1) so shared services such as
MetricsService and a future readiness service no longer collide on the
gateway. Updates the Go control plane and Rust CLI to the new service name
and enables buf/protolint enforcement for the module.
