+++ modules/dscp/cli/build.rs
@@ -1,13 +1,13 @@
 use core::error::Error;
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../controlplane/dscppb/dscp.proto");
+    println!("cargo:rerun-if-changed=../controlplane/dscppb/v1/dscp.proto");
 
     tonic_build::configure()
         .emit_rerun_if_changed(false)
         .build_server(false)
         .message_attribute(".", "#[derive(Serialize)]")
-        .compile_protos(&["dscppb/dscp.proto"], &["../controlplane"])?;
+        .compile_protos(&["dscppb/v1/dscp.proto"], &["../controlplane"])?;
 
     Ok(())
 }
