+++ modules/dscp/controlplane/service_test.go
@@ -12,7 +12,7 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
-	"github.com/yanet-platform/yanet2/modules/dscp/controlplane/dscppb"
+	"github.com/yanet-platform/yanet2/modules/dscp/controlplane/dscppb/v1"
 )
 
 var errBackendFailure = fmt.Errorf("backend failure")
